+++ packages/netlify-cms-backend-git-gateway/src/AuthenticationPage.js
@@ -9,6 +9,7 @@ import {
   colors,
   colorsRaw,
   lengths,
+  zIndex,
 } from 'netlify-cms-ui-legacy';
 
 const LoginButton = styled.button`
@@ -38,7 +39,7 @@ const AuthInput = styled.input`
   margin-top: 6px;
   width: 100%;
   position: relative;
-  z-index: 1;
+  z-index: ${zIndex.zIndex1};
 
   &:focus {
     outline: none;
@@ -164,6 +165,7 @@ export default class GitGatewayAuthenticationPage extends React.Component {
         return (
           <AuthenticationPage
             logoUrl={config.logo_url}
+            siteUrl={config.site_url}
             onLogin={this.handleIdentity}
             renderPageContent={() => (
               <a
@@ -174,14 +176,17 @@ export default class GitGatewayAuthenticationPage extends React.Component {
                 {errors.identity}
               </a>
             )}
+            t={t}
           />
         );
       } else {
         return (
           <AuthenticationPage
             logoUrl={config.logo_url}
+            siteUrl={config.site_url}
             onLogin={this.handleIdentity}
             renderButtonContent={() => t('auth.loginWithNetlifyIdentity')}
+            t={t}
           />
         );
       }