+++ packages/netlify-cms-backend-git-gateway/src/implementation.ts
@@ -133,6 +133,10 @@ export default class GitGateway implements Implementation {
     this.backend = null;
   }
 
+  isGitBackend() {
+    return true;
+  }
+
   requestFunction = (req: ApiRequest) =>
     this.tokenPromise!()
       .then(
@@ -357,7 +361,12 @@ export default class GitGateway implements Implementation {
       { parseText }: { parseText: boolean },
     ) => this.api!.readFile(path, id, { branch, parseText });
 
-    const items = await entriesByFiles([{ path, id }], readFile, 'Git-Gateway');
+    const items = await entriesByFiles(
+      [{ path, id }],
+      readFile,
+      this.api!.readFileMetadata.bind(this.api),
+      'Git-Gateway',
+    );
     const entry = items[0];
     const pointerFile = parsePointerFile(entry.data);
     if (!pointerFile.sha) {