+++ packages/netlify-cms-backend-proxy/src/implementation.ts
@@ -58,6 +58,10 @@ export default class ProxyBackend implements Implementation {
     this.options = options;
   }
 
+  isGitBackend() {
+    return false;
+  }
+
   authComponent() {
     return AuthenticationPage;
   }
@@ -89,7 +93,7 @@ export default class ProxyBackend implements Implementation {
     if (response.ok) {
       return json;
     } else {
-      throw new APIError(json.message, response.status, 'Proxy');
+      throw new APIError(json.error, response.status, 'Proxy');
     }
   }