+++ packages/netlify-cms-core/src/formats/helpers.js
@@ -1,6 +1,6 @@
-export const sortKeys = (sortedKeys = []) => (a, b) => {
-  const idxA = sortedKeys.indexOf(a);
-  const idxB = sortedKeys.indexOf(b);
+export const sortKeys = (sortedKeys = [], selector = a => a) => (a, b) => {
+  const idxA = sortedKeys.indexOf(selector(a));
+  const idxB = sortedKeys.indexOf(selector(b));
   if (idxA === -1 || idxB === -1) return 0;
   if (idxA > idxB) return 1;
   if (idxA < idxB) return -1;