+++ packages/netlify-cms-core/src/lib/__tests__/registry.spec.js
@@ -147,26 +147,4 @@ describe('registry', () => {
       });
     });
   });
-
-  describe('getWidget', () => {
-    it('should throw on missing widget', () => {
-      const { getWidget } = require('../registry');
-
-      expect(() => getWidget('Unknown')).toThrow(
-        new Error(
-          `Could not find widget 'Unknown'. Please make sure the widget name is configured correctly or register it via 'registerwidget'.`,
-        ),
-      );
-    });
-
-    it('should throw on missing widget and suggest lowercase name', () => {
-      const { getWidget, registerWidget } = require('../registry');
-
-      registerWidget('string', {});
-
-      expect(() => getWidget('String')).toThrow(
-        new Error(`Could not find widget 'String'. Did you mean 'string'?`),
-      );
-    });
-  });
 });